+++ src/comps/Filter/Filter.svelte
@@ -1,11 +1,8 @@
 <script lang="ts">
   import FilterTags from './FilterTags.svelte';
-  import FilterColors from './Brickbar/FilterColors.svelte';
-  import FilterParts from './Brickbar/FilterParts.svelte';
-  import FilterPartTypes from './Brickbar/FilterPartTypes.svelte';
   import FilterStates from './FilterStates.svelte';
   import FilterSearch from './FilterSearch.svelte';
-  import { ID_PARTS, lsSiteSettingsKey } from '@interfaces';
+  import { lsSiteSettingsKey } from '@interfaces';
   import { localStore, storedActiveSelection } from '@stores';
   import { stopClick } from '../../utils';
 
@@ -27,21 +24,23 @@
 
     const reasons = [
       'init-tags-url',
+      'init-search-via-url',
       'show-states',
       'show-tags',
       'tooltip-tag-clicked',
       'tag-clicked-close-filter',
+      'remove-all-filters',
       'click-search',
     ];
 
     if (store.site === 'products' && reasons.includes(store.reason)) {
-      if (['show-states', 'show-tags', 'click-search'].includes(store.reason)) {
+      if (['show-states', 'show-tags', 'click-search', 'remove-all-filters'].includes(store.reason)) {
         showFilter = true;
       }
       if (
-        store.reason === 'init-tags-url' ||
-        store.reason === 'tooltip-tag-clicked' ||
-        store.reason === 'tag-clicked-close-filter'
+        ['init-tags-url', 'init-search-via-url', 'tooltip-tag-clicked', 'tag-clicked-close-filter'].includes(
+          store.reason
+        )
       ) {
         showFilter = false;
       }
@@ -62,15 +61,8 @@
 </script>
 
 <details class="card" open={showFilter}>
-  <summary class="large-text" on:click={handleClick}><b>Suche & Filter</b></summary>
-  <FilterSearch {activeSearchString} />
+  <summary class="large-text" on:click={handleClick}><b>Filter & Suche</b></summary>
   <FilterStates {activeStateIds} {activeColorIds} {activePartIds} {activePartTypeIds} {activeSearchString} />
   <FilterTags {activeTagIds} />
-
-  {#if activeTagIds.includes(ID_PARTS) && activeTagIds.length === 1}
-    <h2>Brickbar</h2>
-    <FilterParts {activePartIds} />
-    <FilterColors {activeColorIds} />
-    <FilterPartTypes {activePartTypeIds} />
-  {/if}
+  <FilterSearch {activeSearchString} />
 </details>