+++ src/comps/Product/Products.svelte
@@ -13,6 +13,10 @@
     storedActiveSelection,
     storedActiveProduct,
   } from '@stores';
+  import { ID_MANHATTAN, ID_NETHERLAND, ID_BURG_BLAUSTEIN } from '@interfaces';
+  import Manhattan from '../Specials/Manhattan.svelte';
+  import Netherland from '../Specials/Netherland.svelte';
+  import Blaustein from '../Specials/Blaustein.svelte';
   import Product from './Product.svelte';
   import { handleProductSort } from './sorting';
   import ProductFilter from './ProductFilter.svelte';
@@ -65,6 +69,7 @@
       setUrlParams(urlParam, '');
     }
   });
+
   storedActiveProduct.subscribe(store => {
     // update url
     if ((store.reason === 'open-tooltip' || store.reason === 'click-on-zoom') && store.product.id !== 0) {
@@ -218,6 +223,19 @@
 </script>
 
 <ProductFilter {filteredProducts} />
+
+{#if activeTagIds.includes(ID_MANHATTAN) && activeTagIds.length === 1}
+  <br />
+  <Manhattan />
+{/if}
+{#if activeTagIds.includes(ID_NETHERLAND) && activeTagIds.length === 1}
+  <br />
+  <Netherland />
+{/if}
+{#if activeTagIds.includes(ID_BURG_BLAUSTEIN) && activeTagIds.length === 1}
+  <br />
+  <Blaustein />
+{/if}
 
 <nav class="wrap small-margin no-h-margin">
   {#if extraFilter.parts.count > 0}